requirements.txt
blinker==1.6.2
Bottleneck==1.3.5
click==8.1.3
colorama==0.4.6
Flask==2.3.2
gensim==4.3.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mysql-connector-python==8.0.33
nltk==3.7
numexpr==2.8.4
numpy==1.24.3
pandas==1.5.3
pip==23.0.1
protobuf==3.20.3
PyPDF2==2.10.5
python-dateutil==2.8.2
pytz==2022.7
regex==2022.7.9
scikit-learn==1.2.2
scipy==1.10.1
setuptools==67.8.0
six==1.16.0
smart-open==6.3.0
threadpoolctl==3.1.0
tqdm==4.65.0
Werkzeug==2.3.4
wheel==0.38.4
